Split cli code into diferent files
