+++ libcnmc/cli.py
@@ -8,12 +8,13 @@ from libcnmc.utils import N_PROC
 from libcnmc.core import UpdateCNMCStats, UpdateCINISComptador
 from libcnmc.core.backend import OOOPFactory
 from libcnmc.cir_4_2014 import F1, F1bis, F11
-from libcnmc.cir_3_2015 import F3
 from libcnmc import cir_3_2015
 from libcnmc import cir_4_2015
 from libcnmc.res_4603 import INV
 from libcnmc.res_4603 import CINIMAQ, CINIPOS, CreateCelles, UpdateCINISTrafo
 from libcnmc.res_4603 import UpdateCINISTrams, UpdateCINISCts
+from cli_4666 import cnmc_4666
+from cli_4667 import cnmc_4667
 
 
 @click.group()
@@ -254,41 +255,6 @@ def csv_4667(obj, **kwargs):
     proc.calc()
 
 
-def res_mod(procs, **kwargs):
-    """
-    Generates the file of modificacions from a list of process
-    
-    :param procs:process to generate
-    :type procs: list
-    :param kwargs: Parameters to generate the file, passed to the process
-    :type kwargs: dict(str, str)
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-                    pwd=kwargs['password'], port=kwargs['port'],
-                    uri=kwargs['server'])
-
-    with open(kwargs["output"], "w") as fd:
-        for proc_fnc in procs:
-            temp_fd = tempfile.NamedTemporaryFile()
-            tmp_url = temp_fd.name
-
-            proc = proc_fnc(
-                quiet=kwargs['quiet'],
-                interactive=kwargs['interactive'],
-                output=tmp_url,
-                connection=O,
-                num_proc=kwargs['num_proc'],
-                year=kwargs['year']
-            )
-            proc.calc()
-            with open(tmp_url, "r") as fd_tmp:
-                tmp_data = fd_tmp.read()
-            fd.write(tmp_data)
-
-
 # CSV LAT
 def res_lat(LAT, **kwargs):
     """
@@ -317,68 +283,6 @@ def res_lat(LAT, **kwargs):
         compare_field=kwargs["compare_field"]
     )
     proc.calc()
-
-
-# CSV POS
-def res_pos2(proc1, proc2, **kwargs):
-    """
-
-    :param proc1: generation proces 1
-    :param proc2: generation proces 1
-    :param kwargs:
-    :return:
-    """
-
-    O = OOOPFactory(dbname=kwargs["database"], user=kwargs["user"],
-                    pwd=kwargs["password"], port=kwargs["port"],
-                    uri=kwargs["server"])
-    output = kwargs["output"]
-    temp_fd = tempfile.NamedTemporaryFile()
-
-    tmp_out1 = temp_fd.name
-    temp_fd.close()
-    temp_fd = tempfile.NamedTemporaryFile()
-    tmp_out2 = temp_fd.name
-    temp_fd.close()
-
-    proc = proc1(
-        quiet=kwargs["quiet"],
-        interactive=kwargs["interactive"],
-        output=tmp_out1,
-        connection=O,
-        num_proc=kwargs["num_proc"],
-        codi_r1=kwargs["codi_r1"],
-        year=kwargs["year"],
-        embarrats=kwargs["embarrats"],
-        compare_field=kwargs["compare_field"]
-    )
-    proc.calc()
-
-    proc_2 = proc2(
-        quiet=kwargs["quiet"],
-        interactive=kwargs["interactive"],
-        output=tmp_out2,
-        connection=O,
-        num_proc=kwargs["num_proc"],
-        codi_r1=kwargs["codi_r1"],
-        year=kwargs["year"],
-        embarrats=kwargs["embarrats"],
-        compare_field=kwargs["compare_field"]
-    )
-    proc_2.calc()
-
-    final_out = ""
-    with open(tmp_out1, 'r') as fd1:
-        final_out += fd1.read()
-
-    with open(tmp_out2, 'r') as fd2:
-        final_out += fd2.read()
-
-    with open(output, 'w') as fd_out:
-        fd_out.write(final_out)
-
-    os.unlink(tmp_out1)
-    os.unlink(tmp_out2)
 
 
 @cnmc.command()
@@ -1970,1067 +1874,7 @@ def cir_4_2015_f16(**kwargs):
     )
     proc.calc()
 
-
-def invoke():
-    cnmc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_lat(**kwargs):
-    """
-    Click entry to gnerate the LAT of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-    from libcnmc.res_4131 import LAT, LAT_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(LAT_2015, **kwargs)
-    else:
-        last_year = datetime.now().year - 1
-        kwargs["compare_field"] = "4131_entregada_{}".format(last_year)
-        res_lat(LAT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_lbt(**kwargs):
-    """
-    Click entry to generate the LBT of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-    from libcnmc.res_4131 import LBT, LBT_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(LBT_2015, **kwargs)
-    else:
-        last_year = datetime.now().year-1
-        kwargs["compare_field"] = "4131_entregada_{}".format(last_year)
-        res_lat(LBT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_cts(**kwargs):
-    """
-    Click entry to generate the CTs of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import CTS, CTS_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(CTS_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(CTS, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_sub(**kwargs):
-    """
-    Click entry to generate the SUB fil of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import SUB, SUB_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(SUB_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(SUB, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_pos(**kwargs):
-    """
-    Click entry to generate the POS file of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import POS, POS_2015, POS_INT
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(POS_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_pos2(POS, POS_INT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_maq(**kwargs):
-    """
-    Click entry to generate the MAQ file of 4131
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import MAQ, MAQ_2015
-    if kwargs["year"] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(MAQ_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(MAQ, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_des(**kwargs):
-    """
-    Click entry to generate the DES file
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import DES, DES_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(DES_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(DES, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_fia(**kwargs):
-    """
-    Click entry to generate the fiabilitat file of 4131
-    
-    :param kwargs: Parameters to generate the file
-    :type kwargs: dict(str,str)
-    :return: None
-    :rtype: None
-    """
-    from libcnmc.res_4131 import FIA, FIA_2015
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(FIA_2015, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(FIA, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4131_con(**kwargs):
-    """
-    Click entry to generate the file of condensaodrs of 4131
-    
-    :param kwargs: Params to generate
-    :type kwargs: dict(str, str)
-    :return: None
-    :rtype: None
-    """
-
-    from libcnmc.res_4131 import CON
-    if kwargs['year'] == 2015:
-        kwargs["compare_field"] = "4771_entregada"
-        res_lat(CON, **kwargs)
-    else:
-        kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-        res_lat(CON, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_lat(**kwargs):
-    """
-    Click entry to generate LAT file of 4666
-    
-    :param kwargs: Parameters to generate the LAT file
-    :type kwargs: dict
-    :return: None
-    :rtype: None
-    """
-
-    from libcnmc.res_4666 import LAT
-    last_year = datetime.now().year - 1
-    kwargs["compare_field"] = "4131_entregada_{}".format(last_year)
-    res_lat(LAT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_mod(**kwargs):
-    """
-    Click entry to generate the modification file for 4666
-    
-    :param kwargs: generation arguments
-    :type kwargs: dict
-    :return: None
-    :rtype: None
-    """
-
-    from libcnmc.res_4666 import ModCts, ModFia, ModLat, ModPos
-    from libcnmc.res_4666 import ModLbt, ModMaq
-    procs = [ModCts, ModFia, ModLat, ModLbt, ModMaq, ModPos]
-    res_mod(procs, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_lat(**kwargs):
-    """
-    Click entry to generate the LAT file of 4667
-    
-    :param kwargs: Options to generate the file  
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import LAT
-    csv_4667(LAT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_lbt(**kwargs):
-    """
-    Click entry to generate the LBT of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4666 import LBT
-    last_year = datetime.now().year-1
-    kwargs["compare_field"] = "4131_entregada_{}".format(last_year)
-    res_lat(LBT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_lbt(**kwargs):
-    """
-    Click entry to generate the LBT of 4667 report
-    
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import LBT
-    csv_4667(LBT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_cts(**kwargs):
-    """
-    Click entry to generate the CTs of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4666 import CTS
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(CTS, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_otros(**kwargs):
-    """
-    Click entry to generate the Otros of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import Otros
-    csv_4667(Otros, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_sub(**kwargs):
-    """
-    Click entry to generate the CTs file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4666 import SUB
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(SUB, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_pos(**kwargs):
-    """
-    Click entry to generate the posiciones of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import POS
-    csv_4667(POS, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_pos(**kwargs):
-    """
-    Click entry to generate the POS file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4666 import POS, POS_INT
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_pos2(POS, POS_INT, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_maq(**kwargs):
-    """
-    Click entry to generate the maquines of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import MAQ
-    csv_4667(MAQ, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_maq(**kwargs):
-    """
-    Click entry to generate the MAQ file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import MAQ
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(MAQ, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_des(**kwargs):
-    """
-    Click entry to generate the maquines of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import DES
-    csv_4667(DES, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_des(**kwargs):
-    """
-    Click entry to generate the DES file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4131 import DES
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(DES, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_fia(**kwargs):
-    """
-    Click entry to generate the fiabilitat of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import FIA
-    csv_4667(FIA, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_fia(**kwargs):
-    """
-    Click entry to generate the FIA file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-    from libcnmc.res_4131 import FIA
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(FIA, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4666_con(**kwargs):
-    """
-    Click entry to generate the CON file of 4666
-    
-    :param kwargs: Params to pas to the process
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    from libcnmc.res_4666 import CON
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(CON, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_res(**kwargs):
-    """
-    Click entry to generate the resumen of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import RES
-    csv_4667(RES, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--embarrats/--no-embarrats', default=False,
-              help="Afegir embarrats")
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_466_con(**kwargs):
-    from libcnmc.res_4666 import CON
-    kwargs["compare_field"] = "4131_entregada_{}".format(kwargs["year"])
-    res_lat(CON, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_resccaa(**kwargs):
-    """
-    Click entry to generate the resumen(CCAA) of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import RESCCAA
-    csv_4667(RESCCAA, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_pro(**kwargs):
-    """
-    Click entry to generate the proyectos of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import PRO
-    csv_4667(PRO, **kwargs)
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def res_4667_macro(**kwargs):
-    """
-    Click entry to generate the proyectos of 4667 report
-
-    :param kwargs: Options to generate the file 
-    :return: None 
-    """
-
-    from libcnmc.res_4667 import MACRO
-    csv_4667(MACRO, **kwargs)
+cli = click.CommandCollection(sources=[cnmc, cnmc_4666, cnmc_4667])
 
 if __name__ == '__main__':
-    invoke()
+    cli()
